requirements.txt
hurry.filesize
python-glanceclient
python-keystoneclient
python-novaclient
testtools
